feat: update globals with contexts
