+++ lambda-ui-graphql-relay-ts/src/browser/index.css
@@ -6,15 +6,7 @@ body {
 
 .light {
   background-color: rgb(220, 248, 236);
-  border-radius: 1em;
   border-color: rgb(69, 95, 66);
-  border-style: solid;
-  border-width: 0.12rem;
-  height: 2rem;
-  width: 100%;
-  margin: 1rem;
-  font-size: medium;
-  box-sizing: border-box;
 }
 .light:hover {
   background-color: rgb(171, 247, 218);
@@ -23,17 +15,9 @@ body {
 
 .dark {
   background-color: rgb(52, 55, 54);
-  border-radius: 1em;
   border-color: rgb(69, 95, 66);
-  border-style: solid;
-  border-width: 0.12rem;
-  height: 2rem;
-  width: 100%;
-  margin: 1rem;
-  font-size: medium;
-  box-sizing: border-box;
 }
 .dark:hover {
   background-color: rgb(66, 73, 70);
-  border-color: rgb(73, 169, 63);
+  border-color: rgb(252, 253, 252);
 }
