+++ lambda-ui-graphql-relay-ts/src/components/atoms/buttons/standard/Button.atom.syled.tsx
@@ -23,10 +23,8 @@ const ButtonStyles: React.FC<ButtonStylesProps> = ({ isPrimary = false, isDisabl
 
 const ButtonStyles: React.FC = () => (
   <style>{`
-    .light {
-      // background-color: rgb(220, 248, 236);
+    .button {
       border-radius: 1em;
-      border-color: rgb(69, 95, 66);
       border-style: solid;
       border-width: 0.12rem;
       height: 2rem;
@@ -36,9 +34,7 @@ const ButtonStyles: React.FC = () => (
       box-sizing: border-box;
     }
 
-    .light:hover {
-      // background-color: rgb(171, 247, 218);
-      border-color: rgb(73, 169, 63);
+    .button:hover {
     }
   `}</style>)
