+++ lambda-ui-graphql-relay-ts/src/globals/providers/Providers.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { UserState, ConfigState } from "@states";
+
+function Providers({ children }: any) {
+    return (
+        <ConfigState>
+            <UserState>
+                {children}
+            </UserState>
+        </ConfigState>
+    )
+}
+
+export default Providers
