--- lambda-ui-graphql-relay-ts/src/globals/providers/providers.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import { UserState } from "@states";
-
-function Providers({ children }: any) {
-    return (
-        <UserState>
-            {children}
-        </UserState>
-    )
-}
-
-export default Providers
